+++ include/siddiqsoft/private/restcl_win.hpp
@@ -266,8 +266,8 @@ namespace siddiqsoft
             // typically this is *after* we invoke the callback.
             // The logic here is to invoke the REST request until we get a
             // success response.
-            uint retryCount = 0;
-            uint failCount  = 0;
+            uint16_t retryCount = 0;
+            uint16_t failCount  = 0;
 
             if (arg.retryCounter == 0) {
                 arg.retryCounter = _config[RESTCL_CONFIG_AUTO_REST_RETRY_COUNTER];
@@ -332,10 +332,7 @@ namespace siddiqsoft
                 else {
                     failCount++;
 #if defined(DEBUG)
-                    auto debugLine = std::format("failed#:{}, retry:{}/{}",
-                                                 failCount,
-                                                 arg.retryCounter,
-                                                 MAX_AUTO_RETRY_SEND_LIMIT);
+                    auto debugLine = std::format("failed#:{}, retry:{}/{}", failCount, arg.retryCounter, MAX_AUTO_RETRY_SEND_LIMIT);
                     resp->setHeader("X-restcl-pool-debug", debugLine);
                     resp->setHeader("X-restcl-io-ttx", ttx.lap<std::chrono::milliseconds>());
 #endif
@@ -410,7 +407,7 @@ namespace siddiqsoft
         /// @param callback The method will be async and there will not be a response object returned
         /// @param retryCount When more than 1 and limited to MAX_AUTO_RETRY_SEND_LIMIT it will retry the request until success code
         /// is received. Be careful with this option!
-        basic_restclient& sendAsync(rest_request<>&& req, basic_callbacktype&& callback = {}, uint retryCount = 1) override
+        basic_restclient& sendAsync(rest_request<>&& req, basic_callbacktype&& callback = {}, uint16_t retryCount = 1) override
         {
             if (!isInitialized) throw std::runtime_error("Initialization failed/incomplete!");
